create a setup.py
To facilitate the distribution of the RAI framework, setup.py is added.

In addition, some files are not copied into the package because
some files (e.g., __init__.py) are missing and some others are
non-python files.

Also, certificate_manager code is updated such that the loading
certificate files are agnostic to OSes.